# doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is our api for word trainer app",
    "version": "1.0.1",
    "title": "Word Trainer"
  },
  "host": "thawing-gorge-14317.herokuapp.com",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Gets users list for leaderboard",
        "operationId": "GetUsers",
        "parameters": [
          {
            "name": "rows",
            "in": "query",
            "required": true,
            "type": "integer",
            "description": "rows per page"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer",
            "description": "page sequence number"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserForTable"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get profile info",
        "operationId": "GetProfile",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Error while searching for user"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user info",
        "operationId": "UpdateUser",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user object to register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Error while searching for user"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user profile",
        "operationId": "DeleteUser",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Cannot understand path"
          },
          "404": {
            "description": "No such user"
          },
          "500": {
            "description": "Error happened"
          }
        }
      },
      "post": {
        "summary": "signing up api",
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user object to register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "description": "token, that will be cookie, change that field, dunno how in headers"
            }
          },
          "400": {
            "description": "Error in request"
          }
        }
      }
    },
    "/avatar/": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload an avatar for user",
        "operationId": "UploadAvatar",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "required": true,
            "type": "integer"
          },
          {
            "name": "additionalMetadata",
            "in": "formData",
            "description": "Additional data to pass to server",
            "required": false,
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "description": "file to upload",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error in the path"
          },
          "500": {
            "description": "Error while uploading the file"
          }
        }
      }
    },
...
    "/cards/": {
      "post": {
        "tags": [
          "card"
        ],
        "summary": "Set cards to dictionary from file",
        "operationId": "UploadFile",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "dictionaryId",
            "in": "header",
            "required": true,
            "type": "integer"
          },
          {
            "name": "additionalMetadata",
            "in": "formData",
            "description": "Additional data to pass to server",
            "required": false,
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "description": "file to upload",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error in the path"
          },
          "500": {
            "description": "Error while uploading the file"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserForTable": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "score": {
          "type": "integer",
          "format": "int32",
          "description": "number of points"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "omitempty"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "langID": {
          "type": "integer"
        },
        "pronounceOn": {
          "type": "integer",
          "description": "if pronounciation for user is available"
        },
        "score": {
          "type": "integer",
          "format": "int32",
          "description": "number of points"
        }
      }
    }
  }
}